requirements.txt
einops==0.6.0
matplotlib==3.6.1
nibabel==4.0.0
numpy==1.24.2
pandas==1.5.0
Pillow==9.4.0
scikit_learn==1.2.1
scipy==1.10.1
torch==1.12.1+cu116
torchvision==0.13.1
tqdm==4.64.1
